+++ src/ViewHoldingsRecord.js
@@ -26,6 +26,7 @@ import {
   Button,
   Modal,
   ConfirmationModal,
+  MessageBanner,
 } from '@folio/stripes/components';
 import {
   ViewMetaData,
@@ -54,6 +55,9 @@ import {
   noValue,
   holdingsStatementTypes,
 } from './constants';
+import { WarningMessage } from './components';
+
+import css from './View.css';
 
 class ViewHoldingsRecord extends React.Component {
   static manifest = Object.freeze({
@@ -527,9 +531,22 @@ class ViewHoldingsRecord extends React.Component {
                   </Row>
                   <hr />
                   <AccordionStatus>
-                    <Row end="xs">
-                      <Col xs>
-                        <ExpandAllButton />
+                    <Row className={css.rowMarginBottom}>
+                      <Col xs={11}>
+                        <Row center="xs" middle="xs">
+                          <Col>
+                            <MessageBanner show={Boolean(holdingsRecord.discoverySuppress)} type="warning">
+                              <FormattedMessage id="ui-inventory.warning.holdingsRecord.suppressedFromDiscovery" />
+                            </MessageBanner>
+                          </Col>
+                        </Row>
+                      </Col>
+                      <Col data-test-expand-all xs={1}>
+                        <Row end="xs">
+                          <Col>
+                            <ExpandAllButton />
+                          </Col>
+                        </Row>
                       </Col>
                     </Row>
                     <AccordionSet initialStatus={initialAccordionsState}>
@@ -540,7 +557,7 @@ class ViewHoldingsRecord extends React.Component {
                         <this.cViewMetaData metadata={holdingsRecord.metadata} />
                         <Row>
                           <Col xs={12}>
-                            {holdingsRecord.discoverySuppress && <FormattedMessage id="ui-inventory.discoverySuppress" />}
+                            {holdingsRecord.discoverySuppress && <WarningMessage id="ui-inventory.discoverySuppressed" />}
                           </Col>
                         </Row>
                         <br />