+++ src/views/ItemView.js
@@ -33,6 +33,7 @@ import {
   Icon,
   ConfirmationModal,
   Modal,
+  MessageBanner,
 } from '@folio/stripes/components';
 
 import {
@@ -73,6 +74,7 @@ import {
   actionMenuDisplayPerms,
 } from '../constants';
 import ItemStatus from './ItemStatus';
+import { WarningMessage } from '../components';
 
 export const requestStatusFiltersString = map(requestStatuses, requestStatus => `requestStatus.${requestStatus}`).join(',');
 
@@ -436,7 +438,7 @@ class ItemView extends React.Component {
     const temporaryHoldingsLocation = locationsById[holdingsRecord.temporaryLocationId];
     const tagsEnabled = !tagSettings?.records?.length || tagSettings?.records?.[0]?.value === 'true';
 
-    const requestRecords = (requests || {}).records || [];
+    const requestCount = requests.other?.totalRecords ?? 0;
     const query = location.search ? queryString.parse(location.search) : {};
 
     const requestsUrl = `/requests?filters=${requestStatusFiltersString}&query=${item.id}&sort=Request Date`;
@@ -578,6 +580,7 @@ class ItemView extends React.Component {
       copyNumber: get(item, 'copyNumber', '-'),
       numberOfPieces: get(item, 'numberOfPieces', '-'),
       descriptionOfPieces: get(item, 'descriptionOfPieces', '-'),
+      effectiveShelvingOrder: get(item, 'effectiveShelvingOrder', '-'),
     };
 
     const enumerationData = {
@@ -601,7 +604,7 @@ class ItemView extends React.Component {
       permanentLoanType: get(item, ['permanentLoanType', 'name'], '-'),
       temporaryLoanType: get(item, ['temporaryLoanType', 'name'], '-'),
       itemStatusDate: getDateWithTime(item?.status?.date),
-      requestLink: !isEmpty(requestRecords) ? <Link to={requestsUrl}>{requestRecords.length}</Link> : 0,
+      requestLink: requestCount ? <Link to={requestsUrl}>{requestCount}</Link> : 0,
       borrower: borrowerLink,
       loanDate: openLoan ? getDateWithTime(openLoan.loanDate) : '-',
       dueDate: openLoan ? getDateWithTime(openLoan.dueDate) : '-',
@@ -667,7 +670,7 @@ class ItemView extends React.Component {
     };
 
     const effectiveLocationDisplay = (
-      <Col xs={4} smOffset={0}>
+      <Col xs={2}>
         <KeyValue
           label={<FormattedMessage id="ui-inventory.effectiveLocation" />}
           value={checkIfElementIsEmpty(itemLocation.effectiveLocation)}
@@ -713,7 +716,7 @@ class ItemView extends React.Component {
               >
                 <ModalContent
                   item={item}
-                  itemRequestCount={requestRecords.length}
+                  itemRequestCount={requestCount}
                   status={MISSING}
                   requestsUrl={requestsUrl}
                   onConfirm={this.markItemAsMissing}
@@ -730,7 +733,7 @@ class ItemView extends React.Component {
               >
                 <ModalContent
                   item={item}
-                  itemRequestCount={requestRecords.length}
+                  itemRequestCount={requestCount}
                   status={WITHDRAWN}
                   requestsUrl={requestsUrl}
                   onConfirm={this.markItemAsWithdrawn}
@@ -750,7 +753,7 @@ class ItemView extends React.Component {
               >
                 <ModalContent
                   item={item}
-                  itemRequestCount={requestRecords.length}
+                  itemRequestCount={requestCount}
                   status={itemStatusesMap[selectedItemStatus]}
                   requestsUrl={requestsUrl}
                   onConfirm={() => this.markItemWithStatus(selectedItemStatus)}
@@ -833,14 +836,25 @@ class ItemView extends React.Component {
               <AccordionStatus>
                 <Row>
                   {effectiveLocationDisplay}
-                  <Col xs={8}>
+                  <Col xs={2}>
                     <KeyValue
                       label={<FormattedMessage id="ui-inventory.effectiveCallNumber" />}
                       value={effectiveCallNumber(item)}
                     />
                   </Col>
-                  <Col end="xs">
-                    <ExpandAllButton />
+                  <Col xs={7}>
+                    <Row middle="xs">
+                      <MessageBanner show={Boolean(item.discoverySuppress)} type="warning">
+                        <FormattedMessage id="ui-inventory.warning.item.suppressedFromDiscovery" />
+                      </MessageBanner>
+                    </Row>
+                  </Col>
+                  <Col xs={1}>
+                    <Row end="xs">
+                      <Col>
+                        <ExpandAllButton />
+                      </Col>
+                    </Row>
                   </Col>
                 </Row>
                 <br />
@@ -852,7 +866,7 @@ class ItemView extends React.Component {
                     <ViewMetaData metadata={item.metadata} />
                     <Row>
                       <Col xs={12}>
-                        {item.discoverySuppress && <FormattedMessage id="ui-inventory.discoverySuppress" />}
+                        {item.discoverySuppress && <WarningMessage id="ui-inventory.discoverySuppressed" />}
                       </Col>
                     </Row>
                     {item.discoverySuppress && <br />}
@@ -1371,7 +1385,10 @@ ItemView.propTypes = {
   resources: PropTypes.shape({
     instances1: PropTypes.shape({ records: PropTypes.arrayOf(PropTypes.object) }),
     loanTypes: PropTypes.shape({ records: PropTypes.arrayOf(PropTypes.object) }),
-    requests: PropTypes.shape({ records: PropTypes.arrayOf(PropTypes.object) }),
+    requests: PropTypes.shape({
+      records: PropTypes.arrayOf(PropTypes.object),
+      other: PropTypes.object,
+    }),
     loans: PropTypes.shape({ records: PropTypes.arrayOf(PropTypes.object) }),
     items: PropTypes.shape({ records: PropTypes.arrayOf(PropTypes.object) }),
     holdingsRecords: PropTypes.shape({ records: PropTypes.arrayOf(PropTypes.object) }),